// web/src/types.ts
import { RaRecord, Identifier } from 'react-admin';

export interface Sale extends RaRecord {
    first_name: string;
    last_name: string;
    email: string;
}

export interface Project extends RaRecord {
    name: string;
    env: string;
    infra_arch: number;
    git_url: string;
    tenant_id: string;
    project_id: 0 | 1;
}

export interface Application extends RaRecord {
    app_fullname: string;
    app_shortname: string;
    bizuser_mail: string;
    devuser_mail: string;
    devuser_group: string;
    app_type: string;
    app_description: string;
    app_status: 0 | 1;
    app_id: number;
    project_id: number
}

export interface Company extends RaRecord {
    name: string;
    logo: string;
    sector: string;
    size: 1 | 10 | 50 | 250 | 500;
    linkedIn: string;
    website: string;
    phone_number: string;
    address: string;
    zipcode: string;
    city: string;
    stateAbbr: string;
    nb_contacts: number;
    nb_deals: number;
    sales_id: Identifier;
    created_at: string;
}

export interface Contact extends RaRecord {
    first_name: string;
    last_name: string;
    title: string;
    company_id: Identifier;
    email: string;
    avatar?: string;
    first_seen: string;
    last_seen: string;
    has_newsletter: Boolean;
    tags: Identifier[];
    gender: string;
    sales_id: Identifier;
    nb_notes: number;
}

export interface ContactNote extends RaRecord {
    contact_id: Identifier;
    type: string;
    text: string;
    date: string;
    sales_id: Identifier;
    status: string;
}

export interface Deal extends RaRecord {
    name: string;
    company_id: Identifier;
    contact_ids: Identifier[];
    type: string;
    stage: string;
    description: string;
    amount: number;
    created_at: string;
    updated_at: string;
    sales_id: Identifier;
    index: number;
    nb_notes: number;
}

export interface Tag extends RaRecord {
    name: string;
    color: string;
}